+++ src/main/java/io/github/astrapi69/swing/combobox/model/StringComboBoxModel.java
@@ -1,8 +1,8 @@
 /**
  * The MIT License
- *
+ * <p>
  * Copyright (C) 2015 Asterios Raptis
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
  * "Software"), to deal in the Software without restriction, including
@@ -10,10 +10,10 @@
  * distribute, sublicense, and/or sell copies of the Software, and to
  * permit persons to whom the Software is furnished to do so, subject to
  * the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be
  * included in all copies or substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
  * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -24,15 +24,18 @@
  */
 package io.github.astrapi69.swing.combobox.model;
 
-import java.util.List;
-
 import io.github.astrapi69.collection.array.ArrayExtensions;
+
+import java.util.List;
 
 /**
  * The class {@link StringComboBoxModel} is an implementation of the abstract class
- * {@link AbstractComboBoxModel} for {@link String} values.
+ * {@link GenericComboBoxModel} for {@link String} values.
+ * 
+ * @deprecated use instead the class {@link GenericComboBoxModel}
  */
-public class StringComboBoxModel extends AbstractComboBoxModel<String>
+@Deprecated
+public class StringComboBoxModel extends GenericComboBoxModel<String>
 {
 
 	/** The Constant serialVersionUID. */