+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Version 7.1
 ADDED:
 
 - new constructor with a set from exclude values in class EnumComboBoxModel
+- new class GenericComboBoxModel with generic type value
+- new class GenericMutableComboBoxModel with generic type value
 
 CHANGED:
 
@@ -31,8 +33,8 @@ CHANGED:
 - update of dependency silly-math to new minor version 2.2
 - update of dependency throwable to new minor version 2.3
 - update of test dependency crypt-api to new minor version 8.3
-- update of test dependency file-worker to new minor version 11.1
-- update of test dependency junit-jupiter to new minor version 5.9.0
+- update of test dependency file-worker to new minor version 11.3
+- update of test dependency junit-jupiter to new minor version 5.9.1
 
 Version 7
 -------------